Simpler and more correct iteration algorithm.
The main simplification is replacing the convoluted empty-
subsequence-skipping strategy of finding the last adjacent
separator with a concise search for the next non-separator.
Thanks, @natecook1000.

The main correctness improvement is keeping track of the
number of elements returned from the iterator, so a final
empty subsequence can be returned if necessary in the case
where the base collection ends with a separator.

Added documentation describing the algorithm.

Added test coverage, particularly for
`omittingEmptySubsequences == false`.